+++ src/AirPressureSensor.cc
@@ -46,7 +46,7 @@ class ignition::sensors::AirPressureSensorPrivate
   /// \brief node to create publisher
   public: transport::Node node;
 
-  /// \brief publisher to publish logical camera messages.
+  /// \brief publisher to publish air pressure messages.
   public: transport::Node::Publisher pub;
 
   /// \brief true if Load() has been called and was successful
@@ -57,9 +57,6 @@ class ignition::sensors::AirPressureSensorPrivate
 
   /// \brief Altitude reference, i.e. initial sensor position
   public: double referenceAltitude = 0.0;
-
-  /// \brief Current vertical position of the sensor.
-  public: double verticalPosition = 0.0;
 
   /// \brief Noise added to sensor data
   public: std::map<SensorNoiseType, NoisePtr> noises;
@@ -148,8 +145,7 @@ bool AirPressureSensor::Update(const ignition::common::Time &_now)
   // https://github.com/ethz-asl/rotors_simulator/blob/master/rotors_gazebo_plugins/src/gazebo_pressure_plugin.cpp
   {
     // Get the current height.
-    double height = this->dataPtr->referenceAltitude +
-      this->dataPtr->verticalPosition;
+    double height = this->dataPtr->referenceAltitude + this->Pose().Pos().Z();
 
     // Compute the geopotential height.
     double geoHeight = kEarthRadiusMeters * height /
@@ -201,18 +197,6 @@ void AirPressureSensor::SetReferenceAltitude(double _reference)
 double AirPressureSensor::ReferenceAltitude() const
 {
   return this->dataPtr->referenceAltitude;
-}
-
-//////////////////////////////////////////////////
-void AirPressureSensor::SetVerticalPosition(double _pos)
-{
-  this->dataPtr->verticalPosition = _pos;
-}
-
-//////////////////////////////////////////////////
-double AirPressureSensor::VerticalPosition() const
-{
-  return this->dataPtr->verticalPosition;
 }
 
 IGN_SENSORS_REGISTER_SENSOR(AirPressureSensor)